chore: moves cluster funcs to its own file
