+++ components/modelmeshserving/modelmeshserving.go
@@ -5,7 +5,7 @@ import (
 	"context"
 	dsci "github.com/opendatahub-io/opendatahub-operator/v2/apis/dscinitialization/v1"
 	"github.com/opendatahub-io/opendatahub-operator/v2/components"
-	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/common"
+	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/cluster"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/deploy"
 	operatorv1 "github.com/openshift/api/operator/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -77,7 +77,7 @@ func (m *ModelMeshServing) ReconcileComponent(cli client.Client, owner metav1.Ob
 		if err = m.OverrideManifests(string(platform)); err != nil {
 			return err
 		}
-		err := common.UpdatePodSecurityRolebinding(cli,
+		err := cluster.UpdatePodSecurityRolebinding(cli,
 			[]string{"modelmesh", "modelmesh-controller", "odh-model-controller", "odh-prometheus-operator", "prometheus-custom"},
 			dscispec.ApplicationsNamespace)
 		if err != nil {
